Clean up tests of xprasupport and moves test infrastructure stuff to reahl-dev
